requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
psycopg2-binary==2.9.6
pydantic==2.4.1
pydantic_core==2.10.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
requests==2.31.0
uvicorn==0.23.2
passlib==1.7.4
python-multipart
boto3
python-magic
